Added role-based access control and updated navigation  Implemented role-based access control across various controllers. Added a new field `Role` in `AdminRepository.cs` and updated `AccountController.cs` to handle redirection based on user roles. Added `Authorize` attributes to restrict access based on roles in `AdminController.cs`, `ClassController.cs`, `InstructorController.cs`, `UserController.cs`, and `EnrollmentController.cs`. Moved `MyClasses` method from `UserController.cs` to `AccountController.cs`. Updated `HomeController.cs` to handle unauthorized access. Modified `Program.cs` to include Razor pages and redirect to the unauthorized page for any status code. Updated navigation link in `_Layout.cshtml` and added `Unauthorized.cshtml` to display an unauthorized access page.
