+++ Persistence/Repositories/Entities/AdminRepository.cs
@@ -56,6 +56,7 @@ namespace Persistence.Repositories.ClientRepositories
                     Image = string.IsNullOrEmpty(x.Image) ? string.Empty : string.Concat(AppSetting.DocumentUrl, "\\Assets\\", x.Image),
                     Username = x.Username,
                     Password = x.Password,
+                    Role = "Admin",
                     Status = x.Status,
                     CreatedAt = x.CreatedAt,
                     CreatedBy = x.CreatedBy,
@@ -218,6 +219,7 @@ namespace Persistence.Repositories.ClientRepositories
                     Image = string.IsNullOrEmpty(x.Image) ? string.Empty : string.Concat(AppSetting.DocumentUrl, "\\Assets\\", x.Image),
                     Username = x.Username,
                     Password = x.Password,
+                    Role = "Admin",
                     Status = x.Status,
                     CreatedAt = x.CreatedAt,
                     CreatedBy = x.CreatedBy,
