+++ WebApp/Views/Home/Unauthorized.cshtml
@@ -1,0 +1,25 @@
+﻿@{
+    ViewData["Title"] = "Unauthorized";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<div class="container-fluid bg-primary mb-5">
+    <div class="d-flex flex-column align-items-center justify-content-center" style="min-height: 400px">
+        <h3 class="display-3 font-weight-bold text-white">Not Found</h3>
+        <div class="d-inline-flex text-white">
+            <p class="m-0"><a class="text-white" href="">Home</a></p>
+            <p class="m-0 px-2">/</p>
+            <p class="m-0">Not Found</p>
+        </div>
+    </div>
+</div>
+
+<div class="container-fluid">
+    <div class="row justify-content-center ">
+        <div class="col-12">
+            <div class="text-center">
+                <h1 class="font-weight-bolder text-danger">Page you are looking for is not found!</h1>
+            </div>
+        </div>
+    </div>
+</div>
