+++ WebApp/Views/Shared/Unauthorized.cshtml
@@ -1,0 +1,9 @@
+﻿@{
+    ViewData["Title"] = "Unauthorized";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<div class="text-center">
+    <h1 class="font-weight-bolder">Unauthorized Access</h1>
+    <p>You are not authorized to access this page.</p>
+</div>
