+++ WebApp/Views/Shared/_Layout.cshtml
@@ -82,7 +82,7 @@
                                         <i class="fa fa-user" aria-hidden="true"></i>
                                         My Profile
                                     </a>
-                                    <a href="@Url.Action("MyClasses", "User")" class="dropdown-item">
+                                    <a href="@Url.Action("MyClasses", "Account")" class="dropdown-item">
                                         <i class="fa fa-user" aria-hidden="true"></i>
                                         My Classes
                                     </a>
@@ -219,7 +219,7 @@
              style="border-top: 1px solid rgba(23, 162, 184, 0.2)">
             <p class="m-0 text-center text-white">
                 &copy;
-                <a class="text-primary font-weight-bold" href="#">Your Site Name</a>.
+                <a class="text-primary font-weight-bold" href="#">KidKinder CMS</a>.
                 All Rights Reserved. Designed by
                 <a class="text-primary font-weight-bold" href="https://htmlcodex.com">HTML Codex</a>
             </p>
